[BE] FIX: 라우팅 명세에 맞게 패키지 분리
