+++ backend/src/main/java/proj/pet/member/domain/Member.java
@@ -3,11 +3,13 @@ package proj.pet.member.domain;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import proj.pet.block.domain.Block;
 import proj.pet.category.domain.MemberCategoryFilter;
+import proj.pet.comment.domain.Comment;
+import proj.pet.follow.domain.Follow;
+import proj.pet.reaction.domain.Reaction;
 import proj.pet.report.domain.Report;
-import proj.pet.social.domain.Comment;
-import proj.pet.social.domain.Reaction;
-import proj.pet.social.domain.Scrap;
+import proj.pet.scrap.domain.Scrap;
 
 import java.time.LocalDateTime;
 import java.util.List;
