--- backend/src/main/java/proj/pet/social/service/SocialFacadeService.java
@@ -1,4 +1,0 @@
-package proj.pet.social.service;
-
-public interface SocialFacadeService {
-}
