--- backend/src/main/java/proj/pet/social/service/SocialFacadeServiceImpl.java
@@ -1,9 +1,0 @@
-package proj.pet.social.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class SocialFacadeServiceImpl implements SocialFacadeService {
-}
