--- backend/src/main/java/proj/pet/social/service/SocialService.java
@@ -1,4 +1,0 @@
-package proj.pet.social.service;
-
-public interface SocialService {
-}
